// src/services/models/UserView.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserView = {
  id?: string;
  firstName?: string | null;
  lastName?: string | null;
  fullName?: string | null;
  token?: string | null;
  email?: string | null;
  profilePicture?: string | null;
  phoneNumber?: string | null;
  allowSmsNotification?: boolean;
  allowEmailNotification?: boolean;
  allowPushNotification?: boolean;
};